+++ NEWS.md
@@ -27,6 +27,13 @@ plot(example_A24, lines = "trendA_trisplit")
 
 - `export()`: Dropped experimental warning. reworked the html output. Added a basic output for tau_u. Arguments `caption` and `footnote` allow to specify appearance (if left NA object specific output is generated.). `booktab = TRUE` is now set as a default for kable options.
 
+### Deleted deprecated functions
+
+The following functions were deprected since 2017 and are now removed from scan:
+
+- `makesingleSC()`
+- `makeSCDF()`: Please use `scdf` instead.
+
 # scan 0.52
 
 ## Major changes